% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRR.R
\name{sample.generator}
\alias{sample.generator}
\title{Generating Sample for Sparse Reduced-Rank Regression}
\usage{
sample.generator(n, p, p0, q, r, rho.x, rho.e)
}
\arguments{
\item{n}{sample size.}

\item{p}{column number of prediction matrix X.}

\item{p0}{number of non-zero rows of matrix B.}

\item{q}{column number of response matrix Y.}

\item{r}{rank of coefficient matrix C.}

\item{rho.x}{correlation coefficient of each two columns of prediction matrix X.}

\item{rho.e}{correlation coefficient of each two columns of noise matrix E.}
}
\value{
A list of matrix for the SRRR model, serving as the real model.
}
\description{
Generating samples from multinormal distribution for simulations tesing algorithms of Sparse Reduced-Rank Regression.
}
