% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRR.R
\name{SRRR}
\alias{SRRR}
\title{Fitting Sparse Reduced-Rank Regression Model}
\usage{
SRRR(Y, X, lambda = 0.5, r, method = "sub", weight)
}
\arguments{
\item{Y}{response matrix.}

\item{X}{prediction matrix.}

\item{lambda}{numeric scaling parameter for tuning, with default value 0.5.}

\item{r}{the rank of matrix C.}

\item{method}{a string indicating which method to utilize, where "sub" corresponds to subgradient (default) method. Up to now, this is the only algorithm we have developed, so do not modify the default setting!.}

\item{weight}{a numeric vector of weights of each row of matrix B in the penalty term.}
}
\value{
A list of matrices and iteration. Here we obtain coefficient matrix C and its decomposition matrix B and A.
}
\description{
Estimate the coefficient matrix C by subgradient method.
}
