% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRR.R
\name{SRRR.sim}
\alias{SRRR.sim}
\title{Simulation of Sparse Reduced-Rank Regression}
\usage{
SRRR.sim(n, p, q, p0, r, rho.x, rho.e, runs = 30)
}
\arguments{
\item{n, p, q, p0, r, rho.x, rho.e}{parameters for function "sample.generator"}

\item{runs}{the time we repeat the simulation in order to obtain mean square error, standard error and computation time.}
}
\value{
A list including mean square error, standard error and computation time for functions SRRR and rrpack::srrr, where MSE1, SE1, Time1 correspond to rrpack::srrr, and MSE2, SE2, Time2 correspond to SRRR.
}
\description{
Users may set the size and rank of matrix, namely n,p,q,p0 and r, the correlation coefficients rho.x and rho.e, then the function may compare the results of our function SRRR and function srrr in package "rrpack", which includes mean square error, standard error and computation time.
}
