\name{StatComp22099-package}
\alias{StatComp22099-package}
\docType{package}
\title{
  Final Project of Statistical Computing 22F
}
\description{
  All my homework and several R and Rcpp functions are included.
}
\author{
Ziyuan Lin
Maintainer: Ziyuan Lin <linziyuan@mail.ustc.edu.cn>
}
\keyword{ package } 